added primative block merging
blocks with shared names are now merged. The values of the second
block overwrites the values of the first block. This should be changed
to append the values on property values, and replace for sub-blocks.

Or possibly merge on sub blocks?

eg

@what {
	.color {
		color: blue;
	}
}

@what {
	.color {
		padding: 10px;
	}
}

Will need to recusively merge all the blocks.