Support for font size/line height shorthand
This now works:

	body {
		font: 2em/1em Arial, Verdana;
	}

Note: Putting spaces around the / will cause the parser to do
arithmetic on the units.

The entire thing is seen as a single unit, meaning variables
can't be used inline on either side of the /. But the entire
keyword can be stored in a variable.

Also fixed a bug that threw php error when you left off ; on last line
of file.