changed how functions are evaluated
This mainly applies to lib_ functions, but any function is evalulated to
a type before doing anything. This lets library functions return types
that can be used in equations.

A side effect is that colors are now normalized to the # syntax when
they don't have an alpha component.

lib_ functions that return a raw string will be typed as a keyword, and
those that return a raw php number will be typed as a number. Otherwise
they can return the specific type.

eg, the number type:

	array('number', 1234)

the keyword type

	array('keyword', 'keyword-value')