updated PHP api #247 #246
Added compile, compileFile, checkedCompile, and cachedCompile to
instance. Deprecating parse and passing a file to constructor.

Added setVariables to set initial variables from PHP on the instance.

The static methods ccompile and cexecute now forward to the instance
methods checkedCompile and cachedCompile.

The compile functions should have no configuration arguments. They
should all share the same interface to any compiler settings, which is a
lessc instance.

The deprecated code will still be there, and work the same, I'll just
remove references to it from the documentation.