Make lessc::addParsedFile public
We extend lessc with a custom function that embeds small images in the
generated CSS as base64-encoded data URIs. It'd therefore be useful for us to
tie cache invalidation to the mtimes of these image files by invoking
addParsedFile from the custom function. This patch makes that possible by
changing the visibility of lessc::addParsedFile from protected to public.