Make function helpers public
The set of helper functions that the built-in lib_* functions use to validate
their arguments and generate output should be part of lessphp's public API,
because they are very useful to anyone implementing custom functions.
Accordingly, this patch makes the following protected functions public:

 * assertArgs
 * assertColor
 * assertNumber
 * colorArgs
 * throwError
 * toBool

The patch also makes public the properties $TRUE and $FALSE, since they are
useful for writing custom predicate functions.