Improve handling of /* comments */ when 'preserveComments' is set
When preserveComments is set,
- The top comment in a block should not "jump" outside the block.
- Comments should retain their position relative to the rules that follow them.
- Comments should not be deduplicated.
- When a repeated rule is deduplicated, the comments that precede each
  occurrence should be consolidated above the deduplicated rule.

Because comments and whitespace are handled in tandem, the change impacts
slightly the handling of whitespace. Specifically, 'e(this is simple);' was
previously interpreted as though it were written 'e("this");'. It now causes an
error instead, which is the appropriate behavior (it's how less.js handles it).