tests: Set up phpunit via config file
Use a config file for phpunit and specify the version and script
via composer instead of harcoding for Travis CI.

Also remove the obsolete self-update for composer that was needed
for a bug in an older version of Travis CI workers.

Fix indentation in composer.json (consistently use 4 spaces).

Remove redundant arguments from composer install command.