tests: Apply PSR2 style for 'Generic.WhiteSpace.DisallowTabIndent'
I'm already starting to regret this. Obviously, tabs are the only
correct way of indenting code. But alas, PSR2 disagrees and being
different makes the code harder to deal with for those familiar
with PSR2.

I'd be fine with switching to a different popular style preset,
but we shouldn't have our own mix of PSR2 and something else,
so for let's just go all-in on PSR2.