Added patch for CSS named grids
This adds a patch to lessc->compile that adds a protection to CSS named grids. Once all other parsing is done successfully the protection is removed and the compiled CSS returned.

The protection is a rough patch that removes brackets and replaces them with 3 underscores so [grid-name] becomes ___grid-name___ which does not trigger any parser errors allowing everything to be parsed as normal. If this patch is added into lessphp permanently I would suggest doing 4 underscores for (paranoia) safeties sake or a different pattern that is still safe to parse but less common to be used in someones code.