Patch for CSS named grids.
This patch protects CSS named grids by replacing the brackets before
any parsing or compiling happens. Once all parsing and compiling is
done the brackets are replaced. NOTE: Brackets are changed to 3
underscores which works great but has a small chance of clashing
with someones code, in this case I would suggest changing it to 4
underscores. Who is writing that many in their less/ css?